python main: initial commit
In this commit we add the following files:
- .env
- cli_examples.sh
- main.py
- requirements.txt

The .env file contains the Azure Storage Connection String.

The cli_examples.sh file contains the CLI commands to create a storage account and a container.

The main.py file contains the Python code to interact with the Azure Storage Blob service.